+++ pybedtools/test/test1.py
@@ -1,12 +1,11 @@
 import pybedtools
 import os, difflib, sys
+from nose import with_setup
 from nose.tools import assert_raises, raises
 from pybedtools.helpers import BEDToolsError
 from pybedtools import featurefuncs
+from tfuncs import setup, teardown, testdir, test_tempdir, unwriteable
 
-testdir = os.path.dirname(__file__)
-
-pybedtools.set_tempdir('.')
 
 def fix(x):
     """
@@ -33,6 +32,26 @@ def fix(x):
 # Tabix support tests
 # ----------------------------------------------------------------------------
 
+def make_unwriteable():
+    """
+    Make a directory that cannot be written to and set the pybedtools tempdir
+    to it. This is used to isolate "streaming" tests to ensure they do not
+    write to disk.
+    """
+    if os.path.exists(unwriteable):
+        os.system('rm -rf %s' % unwriteable)
+    os.system('mkdir -p %s' % unwriteable)
+    os.system('chmod -w %s' % unwriteable)
+    pybedtools.set_tempdir(unwriteable)
+
+def cleanup_unwriteable():
+    """
+    Reset to normal tempdir operation....
+    """
+    if os.path.exists(unwriteable):
+        os.system('rm -rf %s' % unwriteable)
+    pybedtools.set_tempdir(test_tempdir)
+
 def test_interval_index():
     """
     supplement to the more general test in test_cbedtools.IntervalTest.testGetItemNegative
@@ -46,6 +65,40 @@ def test_interval_index():
                 'ID=thaliana_1_465_805;match=scaffold_801404.1;rname=thaliana_1_465_805'])
     print iv[4:-3]
     assert iv[4:-3] == ['805', '.']
+
+def test_tuple_creation():
+    # everything as a string
+    t = [
+            ("chr1", "1", "100", "feature1", "0", "+"),
+            ("chr1", "100", "200", "feature2", "0", "+"),
+            ("chr1", "150", "500", "feature3", "0", "-"),
+            ("chr1", "900", "950", "feature4", "0", "+")
+        ]
+    x = pybedtools.BedTool(t).saveas()
+    assert pybedtools.example_bedtool('a.bed') == x
+
+    t = [
+            ("chr1", 1, 100, "feature1", 0, "+"),
+            ("chr1", 100, 200, "feature2", 0, "+"),
+            ("chr1", 150, 500, "feature3", 0, "-"),
+            ("chr1", 900, 950, "feature4", 0, "+")
+        ]
+    x = pybedtools.BedTool(t).saveas()
+    assert pybedtools.example_bedtool('a.bed') == x
+
+    t = [
+            ("chr1", "fake", "gene", "50", "300", ".", "+", ".", "ID=gene1"),
+            ("chr1", "fake", "mRNA", "50", "300", ".", "+", ".", "ID=mRNA1;Parent=gene1;"),
+            ("chr1", "fake", "CDS", "75", "150", ".", "+", ".", "ID=CDS1;Parent=mRNA1;"),
+            ("chr1", "fake", "CDS", "200", "275", ".", "+", ".", "ID=CDS2;Parent=mRNA1;"),
+            ("chr1", "fake", "rRNA", "1200", "1275", ".", "+", ".", "ID=rRNA1;"),]
+    x = pybedtools.BedTool(t).saveas()
+
+    # Make sure that x has actual Intervals and not plain tuples or something
+    assert isinstance(x[0], pybedtools.Interval)
+    assert repr(x[0]) == "Interval(chr1:49-300)"
+    assert x[0]['ID'] == 'gene1'
+
 
 def test_tabix():
     a = pybedtools.example_bedtool('a.bed')
@@ -66,24 +119,18 @@ def test_tabix():
 # ----------------------------------------------------------------------------
 # Streaming and non-file BedTool tests
 # ----------------------------------------------------------------------------
+
+@with_setup(make_unwriteable, cleanup_unwriteable)
 def test_stream():
     """
     Stream and file-based equality, both whole-file and Interval by
     Interval
     """
+    cleanup_unwriteable()
+
     a = pybedtools.example_bedtool('a.bed')
     b = pybedtools.example_bedtool('b.bed')
     c = a.intersect(b)
-
-    # make an unwriteable dir...
-    orig_tempdir = pybedtools.get_tempdir()
-    if os.path.exists('unwriteable'):
-        os.system('rm -rf unwriteable')
-    os.system('mkdir unwriteable')
-    os.system('chmod -w unwriteable')
-
-    # ...set that to the new tempdir
-    pybedtools.set_tempdir('unwriteable')
 
     # this should really not be written anywhere
     d = a.intersect(b, stream=True)
@@ -94,7 +141,7 @@ def test_stream():
     assert d_contents == c_contents
 
     # reconstruct d and check Interval-by-Interval equality
-    pybedtools.set_tempdir('unwriteable')
+    make_unwriteable()
     d = a.intersect(b, stream=True)
 
     for i,j in zip(c, d):
@@ -105,11 +152,11 @@ def test_stream():
     f = pybedtools.example_bedtool('d.gff')
 
     # file-based
-    pybedtools.set_tempdir(orig_tempdir)
+    cleanup_unwriteable()
     g1 = f.intersect(a)
 
     # streaming
-    pybedtools.set_tempdir('unwriteable')
+    make_unwriteable()
     g2 = f.intersect(a, stream=True)
 
     for i,j in zip(g1, g2):
@@ -120,12 +167,9 @@ def test_stream():
     for i in iter(g3):
         print i
 
-    for row in f.cut(range(3), stream=True):
+    for row in a.cut([0, 1, 2, 5], stream=True):
         row[0], row[1], row[2]
-        assert_raises(IndexError, row.__getitem__, 3)
-
-    pybedtools.set_tempdir(orig_tempdir)
-    os.system('rm -fr unwriteable')
+        assert_raises(IndexError, row.__getitem__, 4)
 
 def test_stream_of_stream():
     """
@@ -1125,6 +1169,3 @@ def test_bam_to_fastq():
     y = x.bam_to_fastq(fq=tmpfn)
     assert open(y.fastq).read() == open(pybedtools.example_filename('small.fastq')).read()
 
-def teardown():
-    # always run this!
-    pybedtools.cleanup(remove_all=True)