+++ pybedtools/test/test_contrib.py
@@ -5,6 +5,7 @@ import sys
 import pybedtools
 from pybedtools import Interval
 from pybedtools.contrib import Classifier
+from tfuncs import setup, teardown, testdir, test_tempdir, unwriteable
 
 # model for gdc.
 # chr2L, starts at 1 and spacing is 10bp.